[#301] 1.0.0 dev cnf line refactoring (#319)
* Moving knitting contol logic out of ayab_plugin
