+++ src/main/python/ayab/plugins/ayab_plugin/ayab_control.py
@@ -18,21 +18,36 @@
 #    Andreas Müller, Christian Gerbrandt
 #    https://github.com/AllYarnsAreBeautiful/ayab-desktop
 
+
 from .ayab_communication import AyabCommunication
 from .ayab_communication_mockup import AyabCommunicationMockup
-from . import ayab_image
-import math
+
 import logging
-import os
-from ayab.plugins.knitting_plugin import KnittingPlugin
-from PyQt5 import QtGui, QtWidgets, QtCore
+import pprint
+
 from enum import Enum
 
-from .ayab_options import Ui_DockWidget
-import serial.tools.list_ports
 
-import pprint
-from PIL import ImageOps
+class KnittingState(Enum):
+    NONE = 0
+    SETUP = 1
+    INIT = 2
+    WAIT_FOR_INIT = 3
+    START = 4
+    OPERATE = 5
+    FINISHED = 6
+
+
+class AYABControlKnitResult(Enum):
+    NONE = 0
+    ERROR_SERIAL_PORT = 1
+    CONNECTING_TO_MACHINE = 2
+    WAIT_FOR_INIT = 3
+    ERROR_WRONG_API = 4
+    PLEASE_KNIT = 5
+    DEVICE_NOT_READY = 6
+    FINISHED = 7
+
 
 class KnittingMode(Enum):
     SINGLEBED = 0
@@ -43,367 +58,35 @@ class KnittingMode(Enum):
     CIRCULAR_RIBBER = 4  # Circular Ribber
 
 
-class AyabPluginControl(KnittingPlugin):
-    def onknit(self, e):
-        self.__logger.debug("called onknit on AyabPluginControl")
-
-        if self.conf["continuousReporting"] is True:
-            self.options_ui.tabWidget.setCurrentIndex(1)
-
-        self.__knitImage(self.__image, self.conf)
-        self.finish()
-
-    def onconfigure(self, e):
-        self.__logger.debug("called onconfigure on AYAB Knitting Plugin")
-        # print ', '.join("%s: %s" % item for item in vars(e).items())
-        parent_ui = self.__parent_ui
-
-        # Start to knit with the bottom first
-        pil_image = parent_ui.pil_image.rotate(180)
-        conf = self.get_configuration_from_ui(parent_ui)
-
-        # Mirroring option
-        if conf.get("auto_mirror"):
-            pil_image = ImageOps.mirror(parent_ui.pil_image)
-
-        # TODO: detect if previous conf had the same
-        # image to avoid re-generating.
-
-        try:
-            self.__image = ayab_image.ayabImage(pil_image,
-                                                self.conf["num_colors"])
-        except:
-            self.__notify_user("You need to set an image.", "error")
-            return
-
-        if self.validate_configuration(conf):
-            self.__emit_widget_knitcontrol_enabled(True)
-            self.__emit_button_knit_enabled(True)
-
-            if conf.get("start_needle") and conf.get("stop_needle"):
-                self.__image.setKnitNeedles(conf.get("start_needle"),
-                                            conf.get("stop_needle"))
-            if conf.get("alignment"):
-                self.__image.setImagePosition(conf.get("alignment"))
-
-            self.__image.setStartLine(conf.get("start_line"))
-            self.__emit_progress(
-                conf.get("start_line") + 1, self.__image.imgHeight())
-            self.__emit_color("")
-        else:
-            self.__emit_widget_knitcontrol_enabled(False)
-            self.__emit_button_knit_enabled(False)
-
-        return
-
-    def validate_configuration(self, conf):
-        if conf.get("start_needle") and conf.get("stop_needle"):
-            if conf.get("start_needle") > conf.get("stop_needle"):
-                self.__notify_user("Invalid needle start and end.", "warning")
-                return False
-        if conf.get("start_line") > self.__image.imgHeight():
-            self.__notify_user("Start Line is larger than the image.")
-            return False
-
-        if conf.get("portname") == '':
-            self.__notify_user("Please choose a valid port.")
-            return False
-
-        if conf.get("knitting_mode") == KnittingMode.SINGLEBED.value \
-                and conf.get("num_colors") >= 3:
-            self.__notify_user(
-                "Singlebed knitting currently supports only 2 colors",
-                "warning")
-            return False
-
-        if conf.get("knitting_mode") == KnittingMode.CIRCULAR_RIBBER.value \
-                and conf.get("num_colors") >= 3:
-            self.__notify_user("Circular knitting supports only 2 colors",
-                               "warning")
-            return False
-
-        return True
-
-    def onfinish(self, e):
-        self.__logger.info("Finished Knitting.")
-        self.__close_serial()
-        self.__parent_ui.resetUI()
-
-    def cancel(self):
-        self.__updateNotification("Knitting cancelled")
-        self._knitImage = False
-
-    def __close_serial(self):
-        self.__ayabCom.close_serial()
-
-    def onerror(self, e):
-        # TODO add message info from event
-        self.__logger.error("Error while Knitting.")
-        self.__close_serial()
-
-    def __wait_for_user_action(self, message="", message_type="info"):
-        """
-        Sends the display_blocking_pop_up_signal QtSignal
-        to main GUI thread, blocking it.
-        """
-        self.__parent_ui.signalDisplayBlockingPopUp.emit(message, message_type)
-
-    def __notify_user(self, message="", message_type="info"):
-        """
-        Sends the display_pop_up_signal QtSignal
-        to main GUI thread, not blocking it.
-        """
-        self.__parent_ui.signalDisplayPopUp.emit(message, message_type)
-
-    def __updateNotification(self, message=""):
-        """Sends the signalUpdateNotification signal"""
-        self.__parent_ui.signalUpdateNotification.emit(message)
-
-    def __emit_progress(self, row, total=0, repeats=0):
-        """Sends the updateProgress QtSignal."""
-        self.__parent_ui.signalUpdateProgress.emit(int(row), int(total),
-                                                   int(repeats))
-
-    def __emit_color(self, color):
-        """Sends the updateProgress QtSignal."""
-        self.__parent_ui.signalUpdateColor.emit(color)
-
-    def __emit_status(self, hall_l, hall_r, carriage_type, carriage_position):
-        """Sends the updateStatus QtSignal"""
-        self.__parent_ui.signalUpdateStatus.emit(hall_l, hall_r, carriage_type,
-                                                 carriage_position)
-
-    def __emit_button_knit_enabled(self, enabled):
-        self.__parent_ui.signalUpdateButtonKnitEnabled.emit(enabled)
-
-    def __emit_widget_knitcontrol_enabled(self, enabled):
-        self.__parent_ui.signalUpdateWidgetKnitcontrolEnabled.emit(enabled)
-
-    def __emit_needles(self):
-        """Sends the updateNeedles QtSignal."""
-
-        start_needle_text = self.options_ui.start_needle_edit.value()
-        start_needle_color = self.options_ui.start_needle_color.currentText()
-        start_needle = self.readNeedleSettings(start_needle_color,
-                                               start_needle_text)
-
-        stop_needle_text = self.options_ui.stop_needle_edit.value()
-        stop_needle_color = self.options_ui.stop_needle_color.currentText()
-        stop_needle = self.readNeedleSettings(stop_needle_color,
-                                              stop_needle_text)
-
-        self.__parent_ui.signalUpdateNeedles.emit(start_needle, stop_needle)
-
-    def __emit_alignment(self):
-        """Sends the updateAlignment QtSignal"""
-        alignment_text = self.options_ui.alignment_combo_box.currentText()
-        self.__parent_ui.signalUpdateAlignment.emit(alignment_text)
-
-    def __emit_playsound(self, event):
-        self.__parent_ui.signalPlaysound.emit(event)
-
-    def slotSetImageDimensions(self, width, height):
-        """Called by Main UI on loading of an image to set Start/Stop needle
-    to image width. Updates the maximum value of the Start Line UI element"""
-        left_side = math.trunc(width / 2)
-        self.options_ui.start_needle_edit.setValue(left_side)
-        self.options_ui.stop_needle_edit.setValue(width - left_side)
-        self.options_ui.start_row_edit.setMaximum(height)
-
-    def __onStartLineChanged(self):
-        """ """
-        start_row_edit = self.options_ui.start_row_edit.value()
-        self.__emit_progress(start_row_edit)
-
-    def setup_ui(self, parent_ui):
-        """Sets up UI elements from ayab_options.Ui_DockWidget in parent_ui."""
-        self.set_translator()
-        self.__parent_ui = parent_ui
-        self.options_ui = Ui_DockWidget()
-        self.dock = parent_ui.ui.knitting_options_dock
-        self.options_ui.setupUi(self.dock)
-        self.options_ui.tabWidget.setTabEnabled(1, False)
-        self.options_ui.label_progress.setText("")
-        self.options_ui.label_direction.setText("")
-        self.setup_behaviour_ui()
-
-        # Disable "continuous reporting" checkbox and Status tab for now
-        parent_ui.findChild(QtWidgets.QCheckBox,
-                            "checkBox_ContinuousReporting").setVisible(False)
-        parent_ui.findChild(QtWidgets.QTabWidget, "tabWidget").removeTab(1)
-
-    def set_translator(self):
-        dirname = os.path.dirname(__file__)
-        self.translator = QtCore.QTranslator()
-        self.translator.load(QtCore.QLocale.system(), "ayab_options", ".",
-                             dirname, ".qm")
-        app = QtCore.QCoreApplication.instance()
-        app.installTranslator(self.translator)
-
-    def unset_translator(self):
-        app = QtCore.QCoreApplication.instance()
-        app.removeTranslator(self.translator)
-
-    def populate_ports(self, combo_box=None, port_list=None):
-        if not combo_box:
-            combo_box = self.__parent_ui.findChild(QtWidgets.QComboBox,
-                                                   "serial_port_dropdown")
-        if not port_list:
-            port_list = self.getSerialPorts()
-
-        combo_box.clear()
-
-        def populate(combo_box, port_list):
-            for item in port_list:
-                # TODO: should display the info of the device.
-                combo_box.addItem(item[0])
-
-        populate(combo_box, port_list)
-
-        # Add Simulation item to indicate operation without machine
-        combo_box.addItem("Simulation")
-
-    def setup_behaviour_ui(self):
-        """Connects methods to UI elements."""
-        conf_button = self.options_ui.configure_button
-        conf_button.clicked.connect(self.conf_button_function)
-
-        self.populate_ports()
-        refresh_ports = self.options_ui.refresh_ports_button
-        refresh_ports.clicked.connect(self.populate_ports)
-
-        start_needle_edit = self.options_ui.start_needle_edit
-        start_needle_edit.valueChanged.connect(self.__emit_needles)
-        stop_needle_edit = self.options_ui.stop_needle_edit
-        stop_needle_edit.valueChanged.connect(self.__emit_needles)
-
-        start_needle_color = self.options_ui.start_needle_color
-        start_needle_color.currentIndexChanged.connect(self.__emit_needles)
-        stop_needle_color = self.options_ui.stop_needle_color
-        stop_needle_color.currentIndexChanged.connect(self.__emit_needles)
-
-        alignment_combo_box = self.options_ui.alignment_combo_box
-        alignment_combo_box.currentIndexChanged.connect(self.__emit_alignment)
-
-        start_row_edit = self.options_ui.start_row_edit
-        start_row_edit.valueChanged.connect(self.__onStartLineChanged)
-
-    def conf_button_function(self):
-        self.configure()
-
-    def cleanup_ui(self, parent_ui):
-        """Cleans up UI elements inside knitting option dock."""
-        # dock = parent_ui.knitting_options_dock
-        dock = self.dock
-        cleaner = QtCore.QObjectCleanupHandler()
-        cleaner.add(dock.widget())
-        self.__qw = QtGui.QWidget()
-        dock.setWidget(self.__qw)
-        self.unset_translator()
-
-    def readNeedleSettings(self, color, needle):
-        '''Reads the Needle Settings UI Elements and normalizes'''
-        if (color == "orange"):
-            return 100 - int(needle)
-        elif (color == "green"):
-            return 99 + int(needle)
-
-    def get_configuration_from_ui(self, ui):
-        """Creates a configuration dict from the ui elements.
-
-    Returns:
-      dict: A dict with configuration.
-
-    """
-
-        self.conf = {}
-        continuousReporting = ui.findChild(
-            QtWidgets.QCheckBox, "checkBox_ContinuousReporting").isChecked()
-        if continuousReporting == 1:
-            self.conf["continuousReporting"] = True
-        else:
-            self.conf["continuousReporting"] = False
-
-        color_line_text = ui.findChild(QtWidgets.QSpinBox,
-                                       "color_edit").value()
-        self.conf["num_colors"] = int(color_line_text)
-
-        # Internally, we start counting from zero
-        # (for easier handling of arrays)
-        start_line_text = ui.findChild(QtWidgets.QSpinBox,
-                                       "start_row_edit").value()
-        self.conf["start_line"] = int(start_line_text) - 1
-
-        start_needle_color = ui.findChild(QtWidgets.QComboBox,
-                                          "start_needle_color").currentText()
-        start_needle_text = ui.findChild(QtWidgets.QSpinBox,
-                                         "start_needle_edit").value()
-
-        self.conf["start_needle"] = self.readNeedleSettings(
-            start_needle_color, start_needle_text)
-
-        stop_needle_color = ui.findChild(QtWidgets.QComboBox,
-                                         "stop_needle_color").currentText()
-        stop_needle_text = ui.findChild(QtWidgets.QSpinBox,
-                                        "stop_needle_edit").value()
-
-        self.conf["stop_needle"] = self.readNeedleSettings(
-            stop_needle_color, stop_needle_text)
-
-        alignment_text = ui.findChild(QtWidgets.QComboBox,
-                                      "alignment_combo_box").currentText()
-        self.conf["alignment"] = alignment_text
-
-        self.conf["inf_repeat"] = \
-            int(ui.findChild(QtWidgets.QCheckBox,
-                             "infRepeat_checkbox").isChecked())
-
-        self.conf["auto_mirror"] = \
-            int(ui.findChild(QtWidgets.QCheckBox,
-                             "autoMirror_checkbox").isChecked())
-
-        knitting_mode_index = ui.findChild(QtWidgets.QComboBox,
-                                           "knitting_mode_box").currentIndex()
-        self.conf["knitting_mode"] = knitting_mode_index
-
-        serial_port_text = ui.findChild(QtWidgets.QComboBox,
-                                        "serial_port_dropdown").currentText()
-        self.conf["portname"] = str(serial_port_text)
-
-        # getting file location from textbox
-        filename_text = ui.findChild(QtWidgets.QLineEdit,
-                                     "filename_lineedit").text()
-        self.conf["filename"] = str(filename_text)
-
-        self.__logger.debug(self.conf)
-        # Add more config options.
-        return self.conf
-
-    def getSerialPorts(self):
-        """
-        Returns a list of all USB Serial Ports
-        """
-        return list(serial.tools.list_ports.grep("USB"))
-
+class AYABControl(object):
     def __init__(self):
-        super(AyabPluginControl, self).__init__({})
-
         self.__logger = logging.getLogger(type(self).__name__)
 
-        # Copying from ayab_control
-        self.__API_VERSION = 0x05
+        self._API_VERSION = 0x05
 
-        self.__formerRequest = 0
-        self.__lineBlock = 0
+        self._current_state = KnittingState.SETUP
+        self._progress = dict(current_row=0,
+                              total_rows=0,
+                              repeats=0,
+                              color="",
+                              hall_l=0,
+                              hall_r=0,
+                              carriage_type="",
+                              carriage_position=0)
 
-    def __del__(self):
-        self.__close_serial()
+    def close(self):
+        self.__ayabCom.close_serial()
+
+    def get_progress(self):
+        return self._progress
 
     def __setBit(self, int_type, offset):
+        """ Helper to set a bit within a byte """
         mask = 1 << int(offset)
         return (int_type | mask)
 
     def __setPixel(self, bytearray_, pixel):
+        """ Helper to set a Pixel within a line """
         numByte = int(pixel / 8)
         bytearray_[numByte] = self.__setBit(int(bytearray_[numByte]),
                                             pixel - (8 * numByte))
@@ -452,8 +135,10 @@ class AyabPluginControl(KnittingPlugin):
 
             carriage_position = int(msg[7])
 
-            self.__emit_status(hall_l, hall_r, carriage_type,
-                               carriage_position)
+            self._progress["hall_l"] = hall_l
+            self._progress["hall_r"] = hall_r
+            self._progress["carriage_type"] = carriage_type
+            self._progress["carriage_position"] = carriage_position
 
             return ("indState", msg[1])
 
@@ -753,16 +438,18 @@ class AyabPluginControl(KnittingPlugin):
             self.__logger.debug(msg)
 
             if self.__knitting_mode == KnittingMode.SINGLEBED.value:
-                self.__emit_color("")
+                self._progress["color"] = ""
+                pass
             elif sendBlankLine is True:
                 pass
             else:
-                self.__emit_color(colorNames[color])
+                self._progress["color"] = colorNames[color]
+                pass
 
             # sending line progress to gui
-            self.__emit_progress(imgRow + 1, imgHeight,
-                                 self.__infRepeat_repeats)
-            self.__emit_playsound("nextline")
+            self._progress["current_row"] = imgRow + 1
+            self._progress["total_rows"] = imgHeight
+            self._progress["repeats"] = self.__infRepeat_repeats
 
         else:
             self.__logger.error("requested lineNumber out of range")
@@ -776,97 +463,74 @@ class AyabPluginControl(KnittingPlugin):
         else:
             return 0  # keep knitting
 
-    def __knitImage(self, pImage, pOptions):
-        self.__formerRequest = 0
-        self.__image = pImage
-        self.__startLine = pImage.startLine()
+    def knit(self, pImage, pOptions):
 
-        self.__numColors = pOptions["num_colors"]
-        self.__knitting_mode = pOptions["knitting_mode"]
-        self.__infRepeat = pOptions["inf_repeat"]
-        self.__infRepeat_repeats = 0
+        result = AYABControlKnitResult.NONE
 
-        API_VERSION = self.__API_VERSION
-        curState = 's_init'
-        oldState = 'none'
-
-        if pOptions["portname"] == "Simulation":
-            self.__ayabCom = AyabCommunicationMockup()
-        else:
-            self.__ayabCom = AyabCommunication()
-
-        if not self.__ayabCom.open_serial(pOptions["portname"]):
-            self.__logger.error("Could not open serial port")
-            return
-
-        self._knitImage = True
-        while self._knitImage:
-            # TODO catch keyboard interrupts to abort knitting
-            # TODO: port to state machine or similar.
+        if self._current_state is not KnittingState.SETUP:
             rcvMsg, rcvParam = self.__checkSerial()
-            if curState == 's_init':
-                if rcvMsg == 'cnfInfo':
-                    if rcvParam == API_VERSION:
-                        curState = 's_waitForInit'
-                        self.__updateNotification(
-                            "Please init machine. (Set the carriage to \
-                             mode KC-I or KC-II and move the carriage \
-                             over the left turn mark)."
-                        )
-                    else:
-                        self.__notify_user(
-                            "Wrong Arduino Firmware Version. " +
-                            "Please check if you have flashed " +
-                            "the latest version. (" + str(rcvParam) + "/" +
-                            str(API_VERSION) + ")")
-                        self.__logger.error("wrong API version: " +
-                                            str(rcvParam) + (" ,expected: ") +
-                                            str(API_VERSION))
-                        return
+
+        if self._current_state is KnittingState.SETUP:
+            self.__formerRequest = 0
+            self.__lineBlock = 0
+            self.__image = pImage
+            self.__startLine = pImage.startLine()
+
+            self.__numColors = pOptions["num_colors"]
+            self.__knitting_mode = pOptions["knitting_mode"]
+            self.__infRepeat = pOptions["inf_repeat"]
+
+            self.__infRepeat_repeats = 0
+
+            if pOptions["portname"] == "Simulation":
+                self.__ayabCom = AyabCommunicationMockup()
+            else:
+                self.__ayabCom = AyabCommunication()
+
+            if not self.__ayabCom.open_serial(pOptions["portname"]):
+                self.__logger.error("Could not open serial port")
+                result = AYABControlKnitResult.ERROR_SERIAL_PORT
+
+            self._current_state = KnittingState.INIT
+
+        elif self._current_state == KnittingState.INIT:
+            if rcvMsg == 'cnfInfo':
+                if rcvParam == self._API_VERSION:
+                    self._current_state = KnittingState.WAIT_FOR_INIT
+                    result = AYABControlKnitResult.WAIT_FOR_INIT
                 else:
-                    self.__updateNotification("Connecting to machine...")
-                    self.__ayabCom.req_info()
+                    self.__logger.error("wrong API version: " +
+                                        str(rcvParam) + (" ,expected: ") +
+                                        str(self._API_VERSION))
+                    result = AYABControlKnitResult.ERROR_WRONG_API
+            else:
+                self.__ayabCom.req_info()
+                result = AYABControlKnitResult.CONNECTING_TO_MACHINE
 
-            if curState == 's_waitForInit':
-                if rcvMsg == "indState":
-                    if rcvParam == 1:
-                        curState = 's_start'
-                    else:
-                        self.__logger.debug("init failed")
-
-            if curState == 's_start':
-                if oldState != curState:
+        elif self._current_state == KnittingState.WAIT_FOR_INIT:
+            if rcvMsg == "indState":
+                if rcvParam == 1:
                     self.__ayabCom.req_start(self.__image.knitStartNeedle(),
                                              self.__image.knitStopNeedle(),
                                              pOptions["continuousReporting"])
+                    self._current_state = KnittingState.START
+                else:
+                    self.__logger.debug("init failed")
 
-                if rcvMsg == 'cnfStart':
-                    if rcvParam == 1:
-                        curState = 's_operate'
-                        self.__updateNotification("Please Knit")
-                        self.__emit_playsound("start")
-                    else:
-                        self.__updateNotification()
-                        self.__wait_for_user_action(
-                            "Device not ready, configure and try again.")
-                        self.__logger.error("device not ready")
-                        return
+        elif self._current_state == KnittingState.START:
+            if rcvMsg == 'cnfStart':
+                if rcvParam == 1:
+                    self._current_state = KnittingState.OPERATE
+                    result = AYABControlKnitResult.PLEASE_KNIT
+                else:
+                    self.__logger.error("device not ready")
+                    result = AYABControlKnitResult.DEVICE_NOT_READY
 
-            if curState == 's_operate':
-                if rcvMsg == 'reqLine':
-                    imageFinished = self.__cnfLine(rcvParam)
-                    if imageFinished:
-                        curState = 's_finished'
+        elif self._current_state == KnittingState.OPERATE:
+            if rcvMsg == 'reqLine':
+                imageFinished = self.__cnfLine(rcvParam)
+                if imageFinished:
+                    self._current_state = KnittingState.SETUP
+                    result = AYABControlKnitResult.FINISHED
 
-            if curState == 's_finished':
-                self.__updateNotification("Image transmission finished. "
-                                          "Please knit until you hear the "
-                                          "double beep sound.")
-                self.__emit_playsound("finished")
-                break
-
-            oldState = curState
-
-        self.options_ui.label_carriage.setText("No carriage detected")
-        self.options_ui.tabWidget.setCurrentIndex(0)
-        return
+        return result
